Extract explosion knockback event into its own class
