+++ src/main/java/net/neoforged/neoforge/event/EventHooks.java
@@ -160,6 +160,7 @@ import net.neoforged.neoforge.event.level.BlockGrowFeatureEvent;
 import net.neoforged.neoforge.event.level.ChunkTicketLevelUpdatedEvent;
 import net.neoforged.neoforge.event.level.ChunkWatchEvent;
 import net.neoforged.neoforge.event.level.ExplosionEvent;
+import net.neoforged.neoforge.event.level.ExplosionKnockbackEvent;
 import net.neoforged.neoforge.event.level.LevelEvent;
 import net.neoforged.neoforge.event.level.PistonEvent;
 import net.neoforged.neoforge.event.level.SleepFinishedTimeEvent;
@@ -594,8 +595,18 @@ public class EventHooks {
         NeoForge.EVENT_BUS.post(new ExplosionEvent.Detonate(level, explosion, list));
     }
 
+    /**
+     * To be called when an explosion has calculated the knockback velocity
+     * but has not yet added the knockback to the entity caught in blast.
+     *
+     * @param level             The level that the explosion is in
+     * @param explosion         Explosion that is happening
+     * @param entity            The entity caught in the explosion's blast
+     * @param knockbackVelocity The explosion calculated velocity for the entity
+     * @return The new explosion velocity to add to the entity's existing velocity
+     */
     public static Vec3 onExplosionKnockback(Level level, Explosion explosion, Entity entity, Vec3 knockbackVelocity) {
-        ExplosionEvent.Knockback event = new ExplosionEvent.Knockback(level, explosion, entity, knockbackVelocity);
+        ExplosionKnockbackEvent event = new ExplosionKnockbackEvent(level, explosion, entity, knockbackVelocity);
         NeoForge.EVENT_BUS.post(event);
         return event.getKnockbackVelocity();
     }
